% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aggregate}
\alias{aggregate}
\title{Function \code{aggregate}}
\usage{
aggregate(x, name = "target", aggregator = "list")
}
\arguments{
\item{x}{argument data frame}

\item{name}{name of aggregated object}

\item{aggregator}{function to aggregate the targets}
}
\value{
data frame to aggregate the targets
}
\description{
Aggregate the targets of a previous set of remake commands.
}

