% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(commands, datasets)
}
\arguments{
\item{commands}{output of the \code{\link{commands}} function.}

\item{datasets}{Data frame of commands to generate datasets}
}
\value{
Preprocessed data frame of \code{remake} analysis commands
}
\description{
Produces a data frame of \code{remake} analysis commands.
}
\details{
See https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd
for a package tutorial.
}

