% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_target_names}
\alias{check_target_names}
\title{Function \code{check_target_names}}
\usage{
check_target_names(target_names)
}
\arguments{
\item{target_names}{names of the \code{remake} targets in the workflow}
}
\description{
Checks that the names of targets are legal.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

