% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_stages}
\alias{clean_stages}
\title{Function \code{clean_stages}}
\usage{
clean_stages(stages)
}
\arguments{
\item{stages}{named list of data frames defining the stages of the workflow}
}
\value{
sanitized list of data frames defining the stages of the workflow
}
\description{
Sanitize and check a named list of 
data frames defining the stages of the workflow
}

