% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{commands_batch}
\alias{commands_batch}
\title{Function \code{commands_batch}}
\usage{
commands_batch(x = NULL)
}
\arguments{
\item{x}{named character vector with targets as names and commands as elements}
}
\value{
data frame of remake targets and commands
}
\description{
Similar to \code{\link{commands}} except that commands are
given as a named character vector with targets as names and commands as elements.
\code{\link{commands_string}} is another alternative.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\details{
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{commands}}, \code{\link{commands_string}},
\code{\link{help_remakeGenerator}}
}

