% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{commands_string}
\alias{commands_string}
\title{Function \code{commands_string}}
\usage{
commands_string(...)
}
\arguments{
\item{...}{commands named with their respective targets}
}
\value{
data frame of remake targets and commands
}
\description{
Similar to \code{\link{commands}} except that commands are
parsed as strings rather than symbols. \code{\link{commands_batch}} is
another alternative.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\details{
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{commands}}, \code{\link{commands_batch}},
\code{\link{help_remakeGenerator}}
}

