% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{evaluate}
\alias{evaluate}
\title{Function \code{evaluate}}
\usage{
evaluate(x, wildcard = NULL, values = NULL, expand_x = TRUE)
}
\arguments{
\item{x}{argument data frame}

\item{wildcard}{character string to replace with elements of \code{values}.}

\item{values}{values to replace the wildcard in the remake commands. Must be
the same length as \code{x$command} if \code{expand_x} is \code{TRUE}.}

\item{expand_x}{If \code{TRUE}, loop over \code{values} when evaluating the wildcard,
creating more rows in the output data frame. Otherwise, each occurance of the wildcard
is replaced with the next entry in the \code{values} vector, and the values are recycled.}
}
\value{
an evaluated data frame
}
\description{
Evaluate the wildcards in the commands of a data frame
}

