% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{evaluations}
\alias{evaluations}
\title{Function \code{evaluations}}
\usage{
evaluations(x, rules = NULL, expand = TRUE)
}
\arguments{
\item{x}{argument data frame}

\item{rules}{Named list with wildcards as names and vectors of replacements
as values. See the Details section for more.}

\item{expand}{If \code{TRUE}, loop over the values in \code{rules} when evaluating the wildcards,
creating more rows in the output data frame. Otherwise, each occurance of the wildcard
is replaced with the next entry in the \code{values} vector, and the values are recycled.}
}
\value{
an evaluated data frame
}
\description{
Evaluates multiple wildcards.
}
\details{
Evaluates multiple wildcards by running multiple
calls to \code{evaluate}, with one call for every entry in the 
\code{rules} list. In the calls to \code{evaluate}, the names of 
\code{rules} are plugged into \code{wildcard}, and the elements of
\code{rules} are plugged into \code{values}.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}, \code{\link{evaluate}}
}

