% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_remakeGenerator}
\alias{example_remakeGenerator}
\title{Function \code{example_remakeGenerator}}
\usage{
example_remakeGenerator(example = list_examples_remakeGenerator())
}
\arguments{
\item{example}{name of the example. To see all the available example names, 
run \code{\link{list_examples_remakeGenerator}}.}
}
\description{
Copy a remakeGenerator example to the current working directory.
To see the names of all the examples, run \code{\link{list_examples_remakeGenerator}}.
}
\seealso{
\code{\link{list_examples_remakeGenerator}}, \code{\link{workflow}}
}

