% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor2character}
\alias{factor2character}
\title{Function \code{factor2character}}
\usage{
factor2character(x)
}
\arguments{
\item{x}{a vector, factor or non-factor}
}
\value{
a non-factor vector
}
\description{
Turns a factor into a character and leaves other arguments alone.
}

