% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fake_targets}
\alias{fake_targets}
\title{Function \code{fake_targets}}
\usage{
fake_targets(stages)
}
\arguments{
\item{stages}{named list of data frames defining the stages of the workflow}
}
\value{
\code{YAML}-like list of fake/phony \code{remake} targets
}
\description{
Get a \code{YAML}-like list of fake/phony \code{remake} targets
from a named list of data frames defining the stages of the workflow
}

