% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalize_targets}
\alias{finalize_targets}
\title{Function \code{finalize_targets}}
\usage{
finalize_targets(fake_targets, real_targets)
}
\arguments{
\item{fake_targets}{\code{YAML}-like list of fake/phony \code{remake} targets}

\item{real_targets}{\code{YAML}-like list of real 
(non-fake/non-phony) \code{remake} targets}
}
\value{
\code{YAML}-like list of all \code{remake} targets
}
\description{
Finalize the list of remake targets from 
YAML like lists of fake and real targets.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

