% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help_remakeGenerator}
\alias{help_remakeGenerator}
\title{Function \code{help_remakeGenerator}}
\usage{
help_remakeGenerator()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{workflow}}
}

