% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{list_examples_remakeGenerator}
\alias{list_examples_remakeGenerator}
\title{Function \code{list_examples_remakeGenerator}}
\usage{
list_examples_remakeGenerator()
}
\value{
a names of all the remakeGenerator examples.
}
\description{
Return the names of all the remakeGenerator examples.
}
\seealso{
\code{\link{example_remakeGenerator}}, \code{\link{workflow}}
}

