% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{real_targets}
\alias{real_targets}
\title{Function \code{real_targets}}
\usage{
real_targets(stages)
}
\arguments{
\item{stages}{named list of data frames defining the stages of the workflow}
}
\value{
\code{YAML}-like list of real (non-fake/non-phony) \code{remake} targets
}
\description{
Get a \code{YAML}-like list of real (non-fake/non-phony) \code{remake} 
targets from a named list of data frames defining the stages of the workflow.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

