% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{strings}
\alias{strings}
\title{Function \code{strings}}
\usage{
strings(...)
}
\arguments{
\item{...}{unquoted symbols to turn into character strings.}
}
\value{
a character vector
}
\description{
Turns unquoted symbols into character strings.
}

