% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_random_string}
\alias{unique_random_string}
\title{Function \code{unique_random_string}}
\usage{
unique_random_string(exclude = NULL, n = 30)
}
\arguments{
\item{exclude}{excluded values for the returned character string}

\item{n}{number of characters in the returned character string}
}
\value{
random character string not in \code{exclude}
}
\description{
Generates a random character string that
is different from each of the values in \code{exclude}.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

