% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{workflow}
\alias{workflow}
\title{Function \code{workflow}}
\usage{
workflow(targets = NULL, sources = NULL, packages = NULL,
  remakefile = "remake.yml", makefile = "Makefile", begin = NULL,
  clean = NULL, remake_args = list())
}
\arguments{
\item{targets}{YAML-like list of targets, which you can generate by supplying
data frames of remake commands to the \code{\link{targets}} function.}

\item{sources}{Character vector of R source files}

\item{packages}{Character vector of packages}

\item{remakefile}{Character, name of the \code{remake} file to generate. 
Should be in the current working directory.}

\item{makefile}{Character, name of the Makefile. Should be in the current
working directory. Set to \code{NULL} to suppress the writing of the Makefile.}

\item{begin}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of extra Makefile commands for \code{make clean}.}

\item{remake_args}{Fully-named list of additional arguments to \code{remake::make}.
You cannot set \code{target_names} or \code{remake_file} this way because those
names are reserved.}
}
\description{
Writes the files for a remake workflow.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\details{
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

