% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(commands, datasets)
}
\arguments{
\item{commands}{output of \code{commands(...)}}

\item{datasets}{Data frame of commands to generate datasets}
}
\value{
preprocessed data frame of analysis commands
}
\description{
Helper function for the user. Preprocess a data frame of analysis commands.
}

