% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_commands}
\alias{assert_commands}
\title{Function \code{assert_commands}}
\usage{
assert_commands(x)
}
\arguments{
\item{x}{data frame of remake commands}
}
\description{
Check a data frame of remake commands.
Use the \code{\link{help_remakeGenerator}} function to get more help.
}
\seealso{
\code{\link{help_remakeGenerator}}
}

