% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{commands}
\alias{commands}
\title{Function \code{commands}}
\usage{
commands(...)
}
\arguments{
\item{...}{commands named with their respective targets}
}
\value{
data frame of remake targets and commands
}
\description{
Turn a collection of R expressions into 
a data frame of remake targets and commands.
}

