% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{commands}
\alias{commands}
\title{Function \code{commands}}
\usage{
commands(...)
}
\arguments{
\item{...}{commands named with their respective targets}
}
\value{
data frame of remake targets and commands
}
\description{
Turns a named collection of \code{remake} commands into 
a data frame of \code{remake} targets and \code{remake} commands.
}
\details{
See https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd
for a package tutorial.
}

