% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{evaluate}
\alias{evaluate}
\title{Function \code{evaluate}}
\usage{
evaluate(x, wildcard = NULL, values = NULL, expand = TRUE)
}
\arguments{
\item{x}{argument data frame}

\item{wildcard}{character string to replace with elements of \code{values}.}

\item{values}{values to replace the wildcard in the remake commands. Must be
the same length as \code{x$command} if \code{expand} is \code{TRUE}.}

\item{expand}{If \code{TRUE}, loop over \code{values} when evaluating the wildcard,
creating more rows in the output data frame. Otherwise, each occurance of the wildcard
is replaced with the next entry in the \code{values} vector, and the values are recycled.}
}
\value{
an evaluated data frame
}
\description{
Evaluates the wildcard placeholders of a data frame of \code{remake} commands.
Read https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd for more.
}
\details{
See https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd
for a package tutorial.
}

