% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{example_remakeGenerator}
\alias{example_remakeGenerator}
\title{Function \code{example_remakeGenerator}}
\usage{
example_remakeGenerator(index = 1)
}
\arguments{
\item{index}{index of example (1, 2, etc.)}
}
\description{
Write files to generate an example workflow.
Refer to the generated README.md for further instructions.
}

