% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{strings}
\alias{strings}
\title{Function \code{strings}}
\usage{
strings(...)
}
\arguments{
\item{...}{expressions to turn into characters}
}
\value{
a character vector
}
\description{
Turns expressions into character strings
}

