% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(commands, analyses, datasets, gather = rep("list",
  dim(commands)[1]))
}
\arguments{
\item{commands}{Data frame output of the \code{\link{commands}} function.}

\item{analyses}{Data frame of \code{remake} commands to generate analyses.}

\item{datasets}{Data frame of \code{remake} commands to generate datasets.}

\item{gather}{Character vector, names of functions to gather the summaries.
If not \code{NULL}, length must be the number of rows in the \code{commands}
argument.}
}
\value{
preprocessed data frame of \code{remake} summary commands
}
\description{
Produces a data frame of \code{remake} summary commands.
}
\details{
See https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd
for a package tutorial.
}

