% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(commands, analyses, datasets, gather = rep("list",
  dim(commands)[1]))
}
\arguments{
\item{commands}{Data frame output of \code{commands(...)}}

\item{analyses}{Data frame of commands to generate analyses}

\item{datasets}{Data frame of commands to generate datasets}

\item{gather}{Character vector, names of functions to gather the summaries.
If not \code{NULL}, length must be the number of rows in \code{commands}.}
}
\value{
preprocessed data frame of analysis commands
}
\description{
Helper function for the user. Preprocess a data frame of summary commands.
}

