% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{targets}
\alias{targets}
\title{Function \code{targets}}
\usage{
targets(...)
}
\arguments{
\item{...}{Named collection of data frames of \code{remake} commands.}
}
\value{
YAML-like list of targets.
}
\description{
Puts a named collection of data frames of \code{remake} 
commands all together to make a YAML-like list of targets.
Read https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd for more.
}
\details{
See https://github.com/wlandau/remakeGenerator/blob/master/vignettes/remakeGenerator.Rmd
for a package tutorial.
}

