% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{targets}
\alias{targets}
\title{Function \code{targets}}
\usage{
targets(...)
}
\arguments{
\item{...}{data frames of remake commands}
}
\value{
YAML-like list of targets
}
\description{
Put the data frames of remake commands all together to make a
YAML-like list of targets
}

